Move ForHiveMetastore annotation to thrift package
