+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/thrift/ForHiveMetastore.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive;
+package io.trino.plugin.hive.metastore.thrift;
 
 import com.google.inject.BindingAnnotation;
